Use a random challenge to be signed by the device under verification
Instead of using the UDI we generate and store a random challenge.
This challenge should be signed by the device under verification. The
resulting signature is then signed by the signing server and published
along with the challenge.

Update the README to be consistent with the above and be clearer in
use of terminology.