Rework signing, app binaries handling; break out show-pubkey
In particular, we embed all signal-app binaries found in
internal/appbins/bins/TAG.bin

Support a separate signer-app tag for the vendor signing pubkey. Support
for multiple/rotated vendor signing pubkeys is made be easily
implemented.

The show-pubkey is now a separate command which does not require any
vendor keys for building -- so it's useful for extracting such.