Remove published challenge, sign device public key
Change the way we do signing and verification:

- generate a random challenge every time we want to verify the device
  under verification.

- publish a vendor signature of the device's public key, but not the
  device public key itself.

When verifiying a device we extract the public key from the device and
check the published vendor signature over it. If correct, we proceed.

Then we generate a random challenge and have the device sign it. We
verify the signature against the already verified device public key.
If this signature verifies the TKey has proved that it's in posession
of the corresponding private key and is a genuine TKey.