Rework; adjusting vendor signed msg, adding fwhash, appbin hash, ...
Fork signer-app as verisigner-app and keep it here. This variant can
pull out firmware hash, and also has no touch requirement. It is built
using deps, libs etc from tillitis-key1-apps repo which is expected to
be cloned as a sibling to this repo. Or built by the container build
script, which clones both repos and builds from specific tags. This
script also produces a .sha512-file, and a .deps-file which contains the
specific tag in the apps-repo that was used during the build.

We keep a database of known hardwares (vendorID, productID,
productRevision -- which is encoded into the first word of UDI of TKey
device, called UDI0) and the firmware size and hash that is expected to
be running on each of them.

Both remote-sign and verify commands pull out the UDI and firmware hash
of the TKey "device under verification" and makes sure they are what's
expected.

The message that is signed by the vendor is now UDI(BE)+fwHash+pubkey
(still []byte).

Almost everywhere where previously only the verisigner-app TAG was
specified, there is now also the expected HASH (sha512) of the app
binary. When an app binary is referred to by a TAG, its HASH is also
compared. The only exception is the convenience tool show-pubkey, which
is passed the TAG on the commandline.

Scripts for building app binaries from all (and specific)
verisigner-tags in container is provided. We typically build the
verisigner-app from all tags matching verisigner-vX.Y.Z, embedding all
of them into tkey-verification. From these, the one with the highest
version number is automatically selected as the default app binary --
the one that will be run by the remote-sign command on a new "device
under verification".

The build scripts places all app binaries in internal/appbins/TAG.bin
along with an accompanying TAG.bin.sha512, and as well a .deps-file
which contains the tag of the apps-repo that was used for the
build. Both the .sha512-files and the .deps-files are to be checked in.

Signed-off-by: Daniel Lublin <daniel@lublin.se>