Add show-pubkey command
The show-pubkey command outputs the vendor signing pubkey, the app
tag, and app hash digest to stdout. The output is in the format used
for a line in the vendor-signing-pubkey.txt file, and used to embed
another vendor signing key.

The command takes takes an argument with the path to the devie app
using --app path/to/app.bin