Make vendor signing app digest configurable
Since we verify against all known vendor keys we might as well make
the digest of the app used for vendor signing configurable instead.
It's only used when we do the `serve-siger` command. Let the vendor
decide which of the built-in apps they want to use.

When we list the vendor keys in --built, list all of them, not just
the formerly built-in current vendor key.