Change Sigsum submit key configuration
Add support for many submit keys, configure the active one with
activekey in config. After loading the associated device app verify
that the received public key is the expected one.

For configuration we use a constant string with a simple format like
this:

  name
  ssh key
  tag of app
  app SHA512

We repeat this for all the Sigsum submit keys we allow for
verification.